// src/components/swaps/types/assets.ts
export type AddressOrEth = Address | typeof ETH_ADDRESS;
// ...
export interface ParsedAsset {
  address: AddressOrEth;
  chainId: ChainId;
  chainName: string;
  colors?: {
    primary?: string;
    fallback?: string;
    shadow?: string;
  };
  isNativeAsset: boolean;
  name: string;
  native: {
    price?: {
      change: string;
      amount: number;
      display: string;
    };
  };
  mainnetAddress?: AddressOrEth;
  price?: ZerionAssetPrice;
  symbol: string;
  uniqueId: UniqueId;
  decimals: number;
  icon_url?: string;
  type?: AssetType;
  smallBalance?: boolean;
  standard?: 'erc-721' | 'erc-1155';
  networks?: AssetApiResponse['networks'];
  bridging?: {
    isBridgeable: boolean;
    networks: { [id in ChainId]?: { bridgeable: boolean } };
  };
}

export interface ParsedUserAsset extends ParsedAsset {
  balance: {
    amount: string;
    display: string;
  };
  native: {
    balance: {
      amount: string;
      display: string;
    };
    price?: {
      change: string;
      amount: number;
      display: string;
    };
  };
}

export type SearchAssetWithPrice = SearchAsset & ParsedAsset;
export type ParsedSearchAsset = SearchAsset & ParsedUserAsset;
// ...
export type AssetType = ProtocolType | 'nft';
// ...
export type ProtocolType =
  | 'aave-v2'
  | 'balancer'
  | 'curve'
  | 'compound'
  | 'compound-v3'
  | 'maker'
  | 'one-inch'
  | 'piedao-pool'
  | 'yearn'
  | 'yearn-v2'
  | 'uniswap-v2'
  | 'aave-v3'
  | 'harvest'
  | 'lido'
  | 'uniswap-v3'
  | 'convex'
  | 'convex-frax'
  | 'pancake-swap'
  | 'balancer-v2'
  | 'frax'
  | 'gmx'
  | 'aura'
  | 'pickle'
  | 'yearn-v3'
  | 'venus'
  | 'sushiswap'
  | 'native'
  | 'wrappedNative'
  | 'stablecoin';